ai/chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.84.0
python-dotenv==1.0.1


pymysql
sentence_transformers
faiss-cpu